requirements.txt
aiohttp==3.10.8
azure-cognitiveservices-speech==1.40.0
azure-core==1.31.0
azure-storage-blob==12.23.1
fastapi==0.115.0
ffmpeg-python==0.2.0
moviepy==1.0.3
openai==1.50.2
pydub==0.25.1
pymongo==4.9.1
slack_sdk==3.33.1
uvicorn==0.31.0
